Add form enabling user to submit blog article
“form_for” is a Rails helper method which takes one parameter, in this
case @article and a block with the form fields. The first line
basically says "Create a form for the object named @article, refer to
the form by the name f and add the following elements to the form…"

The “f.label” helper creates an HTML label for a field. This is good
usability practice and will have some other benefits for us later

The “f.text_field” helper creates a single-line text box named title

The “f.text_area” helper creates a multi-line text box named body

The “f.submit” helper creates a button labeled "Create"