Add protection scheme to user registration form
Create a “before_filter” which will run before the “new” and “create”
actions of our authors_controller.rb.

The first line declares that we want to run a before filter named
“zero_authors_or_authenticated” when either the “new” or “create”
methods are accessed. Then we define that filter, checking if there are
either zero registered users OR if there is a user already logged in.
If neither of those is true, we redirect to the root path (our articles
list) and return false. If either one of them is true this filter won’t
do anything, allowing the requested user registration form to be
rendered.